requirements.txt
fastapi==0.110.3
fastapi-standalone-docs==0.1.6
mongomock==4.1.2
mongomock_motor==0.0.30
motor==3.5.1
python-dotenv==1.0.1
uvicorn==0.29.0